Test fixes

+++ app/Http/Controllers/TodoListController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Todo;
 use App\Models\TodoList;
 use Illuminate\Http\Request;
 use App\Http\Requests\StoreTodoListRequest;
@@ -9,6 +10,12 @@ use App\Http\Requests\UpdateTodoListRequest;
 
 class TodoListController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth:api');
+    }
+
+
     /**
      * Display a listing of the resource.
      *
@@ -16,7 +23,19 @@ class TodoListController extends Controller
      */
     public function index(Request $request)
     {
-        return TodoList::paginate();
+
+        $query = $request->user()->todoLists();
+
+        if ($request->has('search')) {
+            $searchTerm = $request->input('search');
+            $query->where(function ($q) use ($searchTerm) {
+                $q->where('title', 'LIKE', "%{$searchTerm}%");
+            });
+        }
+
+        $todoLists = $query->paginate();
+
+        return response()->json(['data' => $todoLists, 'total' => $todoLists->total()]);
     }
 
     /**
@@ -55,6 +74,12 @@ class TodoListController extends Controller
      */
     public function update(UpdateTodoListRequest $request, TodoList $todoList)
     {
+
+        // Ensure the authenticated user owns the TodoList
+        if ($todoList->user_id != $request->user()->id) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
         $todoList->update($request->all());
         return $todoList;
     }
@@ -67,7 +92,17 @@ class TodoListController extends Controller
      */
     public function destroy(TodoList $todoList)
     {
+
+        $user = auth()->user();
+
+        if ($user->id !== $todoList->user_id) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
+
+        Todo::where('list_id', $todoList->id)->delete();
         $todoList->delete();
+
         return response()->json(['message' => 'success'], 200);
     }
 }
